Backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain>=0.3.0,<0.4
langchain-core>=0.3.0,<0.4
langchain-community
langchain-google-genai
langchain-text-splitters
unstructured[pdf,easyocr]
qdrant-client
pytesseract
pdf2image
python-dotenv
google-generativeai
sentence-transformers
torch

# extras required by the provided main.py (do not remove)
nltk
aiofiles
python-magic
pydantic
